Added ability to replace ARC machines, fixes disconnect and reports errors
If the `ForceReplaceAgent` parameter is specified and the agent isn't
connected it will check for any existing agents in the same resource
group and subscription with the same name and if it finds any it will
delete them from Azure using the Arc credentials specified and the Azure
REST API.

Fixes an error in the disconnect step where it was previously supplying
the wrong parameters to `azcmagent`.

Reports any errors to stdout now so DSC reports them to the end user.
Previously they were being swallowed by the process itself and it wasn't
clear why things were failing.